Einführung von Logging mehrere Datenbankkontexte ausarbeitung

+++ BlogApi/Controllers/BlogEntryController.cs
@@ -1,5 +1,6 @@
 ﻿using BlogApi.Model;
 using BlogApi.Model.Entities;
+using BlogApi.Model.Logging;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,51 +9,42 @@ namespace BlogApi.Controllers
 {
     [Route("api/blogentrys")]
     [ApiController]
-    public class BlogEntryController : ControllerBase
+    public class BlogEntryController : Controller
     {
-        private BlogContext db;
-
-        public BlogEntryController(BlogContext context)
-        {
-            db = context;
-        }
+        public BlogEntryController(BlogContext context, ILog logger) :base(context, logger){}
         // GET api/values
         [HttpGet]
         public ActionResult<IEnumerable<BlogEntry>> Get()
         {
-            using (db)
+            return UseDatabaseWithValidModel<IEnumerable<BlogEntry>>(() =>
             {
                 return db.BlogEntrys.ToList();
-            }
+            });
         }
-
         // GET api/values/5
         [HttpGet("{id}")]
         public ActionResult<BlogEntry> Get(int id)
         {
-            using (db)
-            {
-                return db.BlogEntrys.Find(id);
-            }
+            return UseDatabaseWithValidModel(() =>
+                 new ActionResult<BlogEntry>(db.BlogEntrys.Find(id))
+            );
         }
-
         // POST api/values
         [HttpPost]
-        public StatusCodeResult Post([FromBody] BlogEntry value)
+        public ActionResult Post([FromBody] BlogEntry value)
         {
-            using (db)
+            return UseDatabaseWithValidModel(() =>
             {
                 db.BlogEntrys.Add(value);
                 db.SaveChanges();
                 return Ok();
-            }
+            });
         }
-
         // PUT api/values/5
         [HttpPut("{id}")]
-        public StatusCodeResult Put(int id, [FromBody] BlogEntry value)
+        public ActionResult Put(int id, [FromBody] BlogEntry value)
         {
-            using (db)
+            return UseDatabaseWithValidModel(() =>
             {
                 BlogEntry a = db.BlogEntrys.Find(id);
                 a.Categories = value.Categories;
@@ -60,60 +52,76 @@ namespace BlogApi.Controllers
                 a.Subject = value.Subject;
                 db.SaveChanges();
                 return Ok();
-            }
+            });
         }
-
         // DELETE api/values/5
         [HttpDelete("{id}")]
-        public StatusCodeResult Delete(int id)
+        public ActionResult Delete(int id)
         {
-            using (db)
+            return UseDatabaseWithValidModel(() =>
             {
                 BlogEntry a = db.BlogEntrys.Find(id);
                 db.BlogEntrys.Remove(a);
                 db.SaveChanges();
                 return Ok();
-            }
+            });
         }
-
         [HttpPut("/{blogId}/{id}")]
-        public StatusCodeResult Put(int blogId, int id, [FromBody] Comment value)
+        public ActionResult Put(int blogId, int id, [FromBody] Comment value)
         {
-            using (db)
+            return UseDatabaseWithValidModel(() =>
             {
-                BlogEntry a = db.BlogEntrys.Find(blogId);
-                Comment c = a.Comments.Find(x=>x.Id == id);
-                c.Subject = value.Subject;
-                c.Entry = value.Entry;
-                db.SaveChanges();
-                return Ok();
-            }
+                BlogEntry blogEntry = db.BlogEntrys.Find(blogId);
+                if (blogEntry != null)
+                {
+                    Comment c = blogEntry.Comments.Find(x => x.Id == id);
+                    if (c != null)
+                    {
+                        c.Subject = value.Subject;
+                        c.Entry = value.Entry;
+                        db.SaveChanges();
+                        return Ok();
+                    }
+                    else
+                    {
+                        return new StatusCodeResult(902);
+                    }
+                }
+                else
+                {
+                    return new StatusCodeResult(901);
+                }
+            });
         }
 
         [HttpDelete("/{blogId}/{id}")]
-        public StatusCodeResult Delete(int blogId, int id)
+        public ActionResult Delete(int blogId, int id)
         {
-            using (db)
-            {
-                BlogEntry a = db.BlogEntrys.Find(blogId);
-                Comment c = a.Comments.Find(x => x.Id == id);
-                db.Comments.Remove(c);
-                db.SaveChanges();
-                return Ok();
-            }
+           return UseDatabaseWithValidModel(() =>
+                {
+                    BlogEntry blogEntry = db.BlogEntrys.Find(blogId);
+                    if (blogEntry != null)
+                    {
+                        Comment c = blogEntry.Comments.Find(x => x.Id == id);
+                        db.Comments.Remove(c);
+                        db.SaveChanges();
+                        return Ok();
+                    }
+                    else
+                    {
+                        return new StatusCodeResult(901);
+                    }
+                });
         }
-
         [HttpPost("/{blogId}")]
-        public StatusCodeResult Post(int blogId, [FromBody] Comment value)
+        public ActionResult Post(int blogId, [FromBody] Comment value)
         {
-            using (db)
+            return UseDatabaseWithValidModel(() =>
             {
                 db.Comments.Add(value);
                 db.SaveChanges();
                 return Ok();
-            }
+            });
         }
-
-        
     }
 }
